+++ src/ProfilerAwareTrait.php
@@ -8,7 +8,7 @@ trait ProfilerAwareTrait
 {
     protected ?ProfilerInterface $profiler = null;
 
-    public function setProfiler(?ProfilerInterface $profiler): void
+    public function setProfiler(ProfilerInterface $profiler): void
     {
         $this->profiler = $profiler;
     }