+++ tests/ProfilerAwareTraitTest.php
@@ -7,6 +7,7 @@ namespace Yiisoft\Profiler\Tests;
 use PHPUnit\Framework\TestCase;
 use Psr\Log\NullLogger;
 use Yiisoft\Profiler\Profiler;
+use Yiisoft\Profiler\ProfilerAwareInterface;
 use Yiisoft\Profiler\ProfilerAwareTrait;
 use Yiisoft\Profiler\Target\LogTarget;
 use Yiisoft\Profiler\Tests\Logger\ArrayLogger;
@@ -17,7 +18,7 @@ final class ProfilerAwareTraitTest extends TestCase
     {
         $target = new LogTarget(new NullLogger());
         $profiler = new Profiler(new ArrayLogger(), [$target]);
-        $class = new class () {
+        $class = new class () implements ProfilerAwareInterface {
             use ProfilerAwareTrait;
 
             public function getProfiler(): ?Profiler